# src/main/java/com/eagle/bank_api/error/ErrorResponse.java
package com.eagle.bank_api.error;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class ErrorResponse {
    private String message;
}